added rewardValue class

+++ src/main/java/RewardValue.java
@@ -18,6 +18,6 @@ public class RewardValue {
     }
     public double getMilesValue()
     {
-        return milesValue;
+        return (int)milesValue;
     }
 }
